requirements.txt
# Dépendances essentielles pour l'API simple

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Traitement des données
pandas==2.2.2

# Google Sheets et Google Drive
gspread==5.12.0
google-auth==2.23.4
google-api-python-client==2.108.0

# Scraping web
playwright==1.45.0
requests==2.31.0

# Machine Learning
numpy==2.0.2
joblib==1.5.1
sklearn-pandas==2.2.0
scikit-learn==1.6.1